Process page layout TT prior to wrapper TT
Prior to this commit, the order of operations for TT processing was
first to process the beginning of the wrapper, which then internally
call to process the template itself, and then that content would be
passed to a page layout TT if such existed.

With this commit, the order is now: wrapper header first, followed by
the same call to the template itself, but then the returned content from
the template will be passed optionally to a page layout TT prior to
processing any part of the output from wrapper. This allows then for a
project-specific page layout TT to override wrapper defaults.

Additionally, a new default is created: page.layout

This new default is set to "page_layout.html.tt" by default.