Documentation update for import link example
Going back and forth, the desire to organize HTML-based single-file Vue
components as something resulted in the idea of calling them "vuec".

They are technically single-file Vue components, so they could be writen
as *.vue files and put under a vue directory, but they are different
than the typical such files because they're designed to be imported with
the JS import link utility. Thus, the slight group name change.