Remove Bcrypt requirement of salt change
Before, the Bcrypt role would throw an exception if it tried to bcrypt()
without the salt being changed from the default. This is a very good
thing for security purposes but a not so great thing if you have a
newbie user trying to spin up a simple enviornment locally.

This does raise a question about security best practices, but in terms
of this framework, it needs to be assumed to be handled by the app layer
instead of here. In this way, app layer developers can decide how strict
or not strict they want to be with this.