Move dest actions to install Travis phase
Moved the dest init and dest update actions to the Travis CI install
phase because prior to installing modules via cpanm in the install phase
dest does not exist.

Also, instead of installing sass in /usr/local/bin as it should be,
installing it in the project root directory for permissions reasons.