TT "version" variable; page.js array de-specificed
Using the TT version value as a constant with compiled templates means
that even after a restart the constant remains constant, which defeats
the purpose. So it's moved to be a variable now.

While updating the wrapper, expanded the `page.js` array usage to allow
for less-specific naming. The ".js" suffix is optional, and including
the "/js/" path prefix is also optional.