Role::Database class_has attributes become globals
Under the Role::Database is a call to DBIx::Query to create the database
handles, and passed in are handle settings. These DBIx::Query objects
were stored in a class_has attribute. But this is actually undesiredable
behavior at an application level, because each class that's got a role
of Database will re-create the handles. Now, DBIx::Query already caches
the handle and returns it, but DBI itself will override the returned
handle with newly provided handle settings. This all seems fine except
when you have an application that (for example) has DBI AutoCommit on
but turns it off (via begin_work, let's say) to start a transaction.
This is expected to be at the database object level in a model provided
by Role::Model, but it's actually going to result in a shared database
handle having different settings based on where it's calling class is.
Confusing and (likely) never going to be desired behavior. This patch
changes the Role::Database shard caching from class_has to global to fix
this problem. When the shard already exists, Role::Database just returns
it instead of going back to DBIx::Query to pull it from DBI's connection
cache.