Simplify websocket triggers
Previously, there was a semi-complex process for determining what PIDs
should receive an URG to check for websocket data to send. This overly
complex process would work for hypnotoad and morbo (and linda), but it
would fail for tests because the PPID would be incorrect.

Turns out, a much more simple process of looking for the PPID as late as
possible and then maintaining that value works in all contexts.